### 2021-05-06 #3259 @whedon
[PRE REVIEW]: gamut: A Geospatial Analysis of Multisector Urban Teleconnections
**Submitting author:** @KristianNelson (<a href="http://orcid.org/0000-0002-6745-167X">Kristian Nelson</a>)
**Repository:** <a href="https://github.com/IMMM-SFA/gamut" target ="_blank">https://github.com/IMMM-SFA/gamut</a>
**Version:** v0.1.0
