### 2021-05-06 #3259 @kthyng
Hi @KristianNelson and thanks for your submission! I am going to tag this to query the scope, meaning that the editorial board will be pinged to make sure this is in scope for JOSS, the main question being the size since this is relatively small. This should take 1-2 weeks. Thank you for your patience.
