### 2021-05-20 #3259 @kyleniemeyer
@KristianNelson The JOSS editorial board reviewed your submission and found that it does meet our standards to proceed to review—unfortunately, we have low editor availability in this area right now, and so I have to put this on a waitlist until an appropriate editor becomes available.
