+++ man/get_huc8_flows.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/io.R
+\name{get_huc8_flows}
+\alias{get_huc8_flows}
+\title{get_huc8_flows}
+\usage{
+get_huc8_flows()
+}
+\description{
+get_huc8_flows
+}
+\details{
+load huc8 flow table
+}
+\author{
+Kristian Nelson (kristian.nelson@pnnl.gov)
+}
