### 2021-05-28 ba6f1f4
delete raster_to_polygon because it isnt used.
