+++ NAMESPACE
@@ -15,7 +15,6 @@ export(get_zonal_data)
 export(import_raster)
 export(import_shapefile)
 export(mask_raster_to_polygon)
-export(raster_to_polygon)
 export(reclassify_raster)
 export(sup)
 export(suppress_intersect)
@@ -76,7 +75,6 @@ importFrom(sf,st_union)
 importFrom(sf,st_within)
 importFrom(sp,CRS)
 importFrom(sp,SpatialPointsDataFrame)
-importFrom(spex,qm_rasterToPolygons)
 importFrom(stringr,str_remove)
 importFrom(stringr,str_to_title)
 importFrom(tibble,enframe)
