### 2021-06-18 #3259 @KristinaRiemer
Hi @jsta and @fsanchez-trigueros, thank you again for agreeing to review this JOSS submission! The review will happen in the GitHub issue linked to in the above comment, so go over there and check out the guidelines for reviewing and the checklist. Feel free to create issues for the submission author @KristianNelson in the software repo and comment in the review issue. And let me know if you have questions or need anything over there!
