### 2021-06-18 #3259 @danielskatz
ok, it's now fixed - let @openjournals/joss-eics know (in the review issue) if there is any further issue
