### 2021-07-02 #3383 @danielskatz
> I'm not listed as a reviewer on the sidebar of the article proof.

Now you are.  When the PDF you were looking at was generated, you weren't assigned as a reviewer in the issue, and now that you are, I've regenerated the PDF and it shows you.
