### 2021-07-05 IMMM-SFA/gamut#111 @jsta
Consider pinning a version of the sf package
Looks like tests depend on the following `rds` file: `tests/testthat/data/raster_to_polygon.rds`

However, with the latest version of the `sf` package I get the following warning:
> old-style crs object detected; please recreate object with a recent sf::st_crs()

I think this outdated `rds` file is causing the tests in `test-conversion.R` to fail. Seems to me a solution is regenerated the rds file and/or pin an `sf` version in the `DESCRIPTION` file.
