### 2021-07-05 IMMM-SFA/gamut#112 @jsta
Mismatched spelling of UWSCatCH
The Zenodo data product spells the subfolder `water/UWSCatCH` while the code expects `water/UWSCatcH`. Note the lower case vs upper case on the last `C`. This causes `count_watershed_teleconnections` to error.
