### 2021-07-05 5abfa44
Fix UWSCatCH mispelling, closes #112
