### 2021-07-05 #3383 @jsta
Here is my review @KristianNelson @KristinaRiemer :

## Checklist

### Documentation

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I don't see guidelines for 1 and 2. Often `R` packages use a `BugReports` field in the DESCRIPTION for 2 and/or a `CONTRIBUTING` file for 1.

### Software Paper

> Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

I feel like the aim of the paper is a bit muddled. Much of the text and the title suggest that the aim of the paper is to report the results of an analysis but I think JOSS papers are typically focused on describing software tools (not reporting on an analysis). Perhaps the paper could include more text describing how `gamut` works beyond using "watershed delineations ... to mask ... geospatial land use layers"? How does this happen? What kind of decisions (if any) does `gamut` make behind the scenes to accomplish this task? Do some datasets require special handling? Are there any pieces of `gamut` as a software package that might be useful in other MSD projects?

> Do the authors describe how this software compares to other commonly-used packages?

Are there any other software packages for MSD analysis? If not, might be good to say this in the paper.

## Curiosity questions and nice to haves (not required)

How are Los Angeles and other large cities that draw water from outside their watershed handled by `gamut`?

I wish the function documentation was a little more informative. https://immm-sfa.github.io/gamut/reference/index.html just repeats the function name in function descriptions.

How can we make a map of a given watershed like in the paper figure? Is there a function for that?

I think some of the `clisymbols` and `crayon` functionality is dependent on RStudio? If so, might be good to list that as a suggested IDE is some capacity. I'm not seeing them when I use [radian](https://github.com/randy3k/radian) and VSCode.

## Issues

https://github.com/IMMM-SFA/gamut/issues/112, https://github.com/IMMM-SFA/gamut/issues/111
