### 2021-07-07 #3383 @KristianNelson
@jsta Thank you for your feedback! I will add in those necessary guidelines and I do agree we need to clear up the aim of the paper so that it focuses more on how the package works. You make some really good points that I will consider in the edits.

To answer your first question, we utilize the Urban Water Blueprint to know where cities are getting their drinking water. Los Angeles has multiple watersheds all throughout California, but also pulls in water from the Colorado River Basin using the Colorado River Aqueduct which starts at Lake Havasu. We analyze the hydrological statistics upstream of Lake Havasu and include them in the results for Los Angeles. It is important to note that `gamut` also considers the contribution of watersheds to each city. So watersheds that contribute less to a cities overall drinking water supply have less of an impact on the final statistics.

For question 2, I will look into making the function documentation more informative for users.

We did have a function called `plot_watershed` where a specific city could be named and it would produce a map of their watersheds. We removed it prior to submission, but we can definitely place it back in if you think that is a helpful addition to the package.

Thank you again for the feedback! It is greatly appreciated.
