### 2021-07-07 30d306e
add drought calc
