+++ R/count_teleconnections.R
@@ -15,6 +15,7 @@
 count_watershed_teleconnections <- function(data_dir,
                                             cities = NULL,
                                             run_all = TRUE,
+                                            drought_calc = FALSE,
                                             file_paths = c(
                                               watersheds = "water/CWM_v2_2/World_Watershed8.shp",
                                               withdrawal = "water/CWM_v2_2/Snapped_Withdrawal_Points.shp",
@@ -37,6 +38,7 @@ count_watershed_teleconnections <- function(data_dir,
   sup(count_watershed_data(data_dir = data_dir,
                        cities = cities,
                        run_all = run_all,
+                       drought_calc = drought_calc,
                        file_paths = file_paths)) ->
     watershed_data
 
@@ -321,7 +323,7 @@ count_watershed_teleconnections <- function(data_dir,
             select(x = one_of(metric)) %>%
             filter(!is.na(x)) %>% .[["x"]] %>% mean()
         }
-
+        #FIX THIS
         get_max_runoff_conc <- function(metric){
           runoff_contaminant_and_contributions %>%
             select(x = one_of(metric)) %>%
@@ -484,68 +486,69 @@ count_watershed_teleconnections <- function(data_dir,
         # output
         return(
           tibble(city,
-                 city_population,
-                 n_watersheds,
-                 n_other_cities,
-                 dependent_city_pop,
-                 watershed_area_sqkm,
-                 storage_BCM,
-                 yield_BCM,
-                 irr_cons_BCM,
-                 n_climate_zones,
-                 #n_transfers_in,
-                 #n_transfers_out,
-                 #n_transfers_within,
-                 n_hydro_plants,
-                 n_thermal_plants,
-                 n_fac_agcrop,
-                 n_fac_aglivestock,
-                 n_fac_cnsmnf,
-                 n_fac_mining,
-                 n_fac_oilgas,
-                 n_fac_total,
-                 hydro_gen_MWh,
-                 thermal_gen_MWh,
-                 thermal_cons_BCM,
-                 thermal_with_BCM,
-                 n_utilities,
-                 n_ba,
-                 n_crop_classes,
-                 cropland_fraction,
-                 #cropland_runoff_percent,
-                 developed_fraction,
-                 #developed_runoff_percent,
-                 ag_runoff_max,
-                 ag_runoff_av_exgw,
-                 ag_runoff_av,
-                 dev_runoff_max,
-                 dev_runoff_av_exgw,
-                 dev_runoff_av,
-                 np_runoff_max,
-                 np_runoff_av_exgw,
-                 np_runoff_av_exgw_unweighted,
-                 np_runoff_av,
-                 n_economic_sectors,
-                 max_withdr_dist_km,
-                 avg_withdr_dis_km,
-                 n_treatment_plants,
-                 watershed_pop,
-                 pop_cons_m3sec,
-                 av_fl_sur_conc_pct = surface_water_concentrations[["average historical flow"]],
-                 av_fl_sur_conc_pct_unweighted = surface_water_concentrations_unweighted[["average historical flow"]],
-                 #dr_fl_sur_conc_pct = surface_water_concentrations[["driest month average flow"]],
-                 av_ro_sur_conc_pct = surface_water_concentrations[["average historical runoff"]],
-                 #dr_ro_sur_conc_pct = surface_water_concentrations[["driest month average runoff"]],
+                 #city_population,
+                 #n_watersheds,
+                 #n_other_cities,
+                 #dependent_city_pop,
+                 #watershed_area_sqkm,
+                 #storage_BCM,
+                 #yield_BCM,
+                 #irr_cons_BCM,
[…]
