+++ man/count_watershed_teleconnections.Rd
@@ -8,6 +8,7 @@ count_watershed_teleconnections(
   data_dir,
   cities = NULL,
   run_all = TRUE,
+  drought_calc = FALSE,
   file_paths = c(watersheds = "water/CWM_v2_2/World_Watershed8.shp", withdrawal =
     "water/CWM_v2_2/Snapped_Withdrawal_Points.shp", citypoint =
     "water/CWM_v2_2/City_Centroid.shp", powerplants =
