### 2021-07-07 bdbdf4c
correct typos (joss edits)
