### 2021-08-17 #3383 @KristinaRiemer
Hi everyone! @KristianNelson have you had a chance to add contribution guidelines or update the paper using the suggestions from @jsta yet? And @fsanchez-trigueros have you had a chance to start your review of this yet? Please let me know how I can help or if there are any questions!
