### 2021-08-18 #3383 @KristianNelson
@KristinaRiemer I am currently working on a branch that is addressing @jsta first round of edits! This will include fixes to code errors, paper edits, and a `CONTRIBUTING.md` file for bug reports and contributing guidelines. I will merge this one and make a new branch for any further reviews.
