+++ paper/paper.md
@@ -1,42 +1,45 @@
 ---
-title: 'gamut: A Geospatial Analysis of Multisector Urban Teleconnections'
+title: 'gamut: A Geospatial R Package to Analyze Multisector Urban Teleconnections'
 tags:
-  - R
-  - Multisector Dynamics
-  - Water
-  - Energy
-  - Land
-  - Urban
-  - Geospatial
+- R
+- Multisector Dynamics
+- Water
+- Energy
+- Land
+- Urban
+- Geospatial
+date: "27 August 2021"
+output:
+  html_document: 
+    df_print: paged
 authors:
-  - name: Kristian D. Nelson
-    orcid: 0000-0002-6745-167X
-    affiliation: 1
-  - name: Sean W. Turner
-    orcid: 0000-0003-4400-9800
-    affiliation: 1
-  - name: Chris Vernon
-    orcid: 0000-0002-3406-6214
-    affiliation: 1
-  - name: Jennie Rice
-    orcid: 0000-0002-7833-9456
-    affiliation: 1
-affiliations:
- - name: Pacific Northwest National Laboratory
-   index: 1
-date: 26 April 2020
+- name: Kristian D. Nelson
+  orcid: 0000-0002-6745-167X
+  affiliation: 1
+- name: Sean W. Turner
+  orcid: 0000-0003-4400-9800
+  affiliation: 1
+- name: Chris Vernon
+  orcid: 0000-0002-3406-6214
+  affiliation: 1
+- name: Jennie Rice
+  orcid: 0000-0002-7833-9456
+  affiliation: 1
 bibliography: paper.bib
+affiliations:
+- name: Pacific Northwest National Laboratory
+  index: 1
 ---
 
 ### Summary
 
 Most cities in the United States withdraw surface water to meet public water supply needs. The lands on which this water is generated are often developed for human activities&mdash;such as agriculture, mining, and industry&mdash;that may compete for water resources or contaminate water supplies. Cities are thereby connected to other sectors through their water supply catchments. This connection is an example of an multisectoral urban teleconnection. The Geospatial Analytics for Multisectoral Urban Teleconnections (`gamut`) package provides national-scale information on these teleconnections by combining land use data with hydrological analysis to characterize urban source watershed human interactions across the conterminous United States (Figure 1).
 
-![The `gamut` package analyzes urban cities and their watersheds all across the conterminous U.S. As shown in the figure, it can look at characteristics like land use and facility operations inside watershed boundaries.](gamut_figure.png){ width=95% }
+![The `gamut` package analyzes urban cities and their watersheds all across the conterminous U.S. As shown in the figure, it can look at characteristics like land use inside watershed boundaries.](gamut_figure.png){ width=95% }
 
-The `gamut` package computes dozens of city-level metrics that inform on the geographical nature of surface water supply catchments and the presence, intensity, and impact of human activities in those catchments. Each city’s watersheds are based on the Urban Water Blueprint [@McDonald:2014], which is enhanced with source contribution estimates as well as river flow and high-resolution runoff [@Nelson:2021]. Watershed delineations are used to mask several geospatial land use layers relating to electricity generation, agriculture, industry and other economic developments, and water infrastructure (dams, reservoirs, aqueducts). These geospatial input layers have been combined into an open-source dataset and can accessed [here](https://zenodo.org/record/4662993#.YJB9ZLVKhPY) [@Nelson2:2021].
[…]
