### 2021-09-14 #3383 @fsanchez-trigueros
Here is my review @KristianNelson @KristinaRiemer:

### Overall comment

Thanks to its tools and the data sets (https://zenodo.org/record/4662993/files/Geospatial_Input_Datasets.zip), gamut seems useful to streamline current environmental and socioeconomic analysis of main CONUS cities based on watershed location, but the paper may still need revisions to contextualize its basic terminology, describe the state of the art in MSD and urban teleconnections, illustrate real case studies that demonstrate the power of the method and the gamut tools, clarify the possibility of further software and data expansions (for now it can only be used with a limited set of geodata in CONUS, how might it be updated to be more versatile and useable in other regions or with custom data?), or comment on its capacity for the open-source community to participate in its development

### Checklist

In the comments below, I bundle checklist questions first (in bold) and then list comments that relate to the bundle:

**Installation: Does installation proceed as outlined in the documentation?**

**Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.**

- devtools is necessary for installation and it is explained in the package, but I recommend adding the lines showing the use of install.packages() and library() to get devtools to work, in case the user does not have devtools installed and is new to R.

- In addition to devtools, my R installation required Rtools but I could not install it using install-packages() from my local up-to-date R 4.1 installation (Rtools is not listed in CRAN for 4.1). I was able to get it to work by downloading the install file at the Rtools package website, then installing it via Windows, and modifying the PATH file to access make and like functions. It might be useful to explain this in the instructions, as users might not have Rtools installed and activated.

- Remote installation from GitHub did not work for me. I got the below error from R:

> devtools::install_github('IMMM-SFA/gamut')

Downloading GitHub repo IMMM-SFA/gamut@HEAD
Error in utils::download.file(url, path, method = method, quiet = quiet,  :
  download from 'https://api.github.com/repos/IMMM-SFA/gamut/tarball/HEAD' failed

So I saved the gz file on my machine from https://api.github.com/repos/IMMM-SFA/gamut/tarball/HEAD, then I run install_local() with my local copy of the source code:

> install_local('C:\\Users\\fsanc\\Downloads\\IMMM-SFA-gamut-v0.1.0-340-g0db7d24.tar.gz')

I know how to troubleshoot this in R but potential users may not, and I didn’t get to find instructions in the gamut materials to help users navigate this potential issue when unable  to install remotely.

**Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).**

**References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?**

**State of the field: Do the authors describe how this software compares to other commonly-used packages?**

- In the Description section (paper, vignette), the notion of “urban teleconnection” is briefly explained but no citation about it is provided. For most potential users, it might not be clear what science lies behind the concept (Is it an adaptation of climate teleconnections? Who coined the term? What publications talk about it?) or whether it aims to be a paradigm coined by the authors of the gamut package [Following NOAA, a teleconnection is a “recurring and persistent, large-scale pattern of pressure and circulation anomalies that spans vast geographical areas.” That is, atmospheric patterns that repeat themselves and relate to each other at large distances (e.g., the El Niño Oscillation).]

- Urban teleconnections seem to share some semantics with atmospheric teleconnections, but it is not clear what science and practice it is based on, so additional bibliographical references seem necessary to explain the motivations, and previous and potential use cases of this tool.

- In sum, I recommend extending the literature review relative to  MSD and urban teleconnections, as well as references to case studies that illustrate the methods enabled by this too. This is missing in the paper but should help attract the interest of potential package users. Even though the tool is useful to advance spatial analysis in CONUS, the current draft of the paper may still be failing short to be able to brand it appropriately among potential users.

### Other comments

- The data ZIP file is 1.2 Gb. I wonder if there is a repository different from Zenodo or an alternative way of compressing or organizing the data files. Downloading from the Zenodo repository ranged from 40KB to 140KB per sec, taking about 6 hours on a business day. I run into failed attempts too after a few hours downloading. This might have a negative impact to attract potential users if these issues persist.

- In https://github.com/IMMM-SFA/gamut, the example showing results of a call of count_watershed_teleconnections() with Portland, OR etc. contains more columns than the table below describing the meaning of each output variable, i.e., some columns from the results (e.g., **n_transfers_in**) are not explained in README and the vignette.
