+++ .github/workflows/test_coverage.yml
@@ -4,36 +4,35 @@ name: build
 
 jobs:
   build:
-    runs-on: macOS-latest
-    strategy:
-      matrix:
-        r: ['3.6']
+    runs-on: ubuntu-latest
+
+    env:
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
 
     steps:
       - uses: actions/checkout@v1
       - uses: r-lib/actions/setup-r@master
         with:
-          r-version: ${{ matrix.r }}
+          r-version: '3.6'
       - uses: r-lib/actions/setup-pandoc@master
       - uses: r-lib/actions/setup-tinytex@master
 
-      - name: Brew and macOS config
-        if: runner.os == 'macOS'
+      - name: Cache R packages
+        uses: actions/cache@v1
+        with:
+          path: ${{ env.R_LIBS_USER }}
+          key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }}
+          restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-
+
+      - name: Ubuntu config
         run: |
-          brew install pkg-config \
-            imagemagick@6 \
-            gdal
+          sudo add-apt-repository ppa:ubuntugis/ppa
+          sudo apt-get update
+          sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev libudunits2-dev libgdal-dev libgeos-dev libproj-dev libmagick++-dev libavfilter-dev
 
-      - name: Install dependencies
-        env:
-          R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-        run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'covr'))" -e "remotes::install_deps(dependencies = TRUE)"
-
-      - name: Check
-        run: Rscript -e "rcmdcheck::rcmdcheck(error_on = 'error')"
 
       - name: Test coverage
-        if: matrix.r == '3.6'
         run: |
-          Rscript -e 'remotes::install_github("r-lib/covr@gh-actions")'
-          Rscript -e 'covr::codecov(token = "${{secrets.CODECOV_TOKEN}}")'
+          remotes::install_cran("covr")
+          covr::codecov(token = "${{secrets.CODECOV_TOKEN}}")
+        shell: Rscript {0}
