### 2021-09-23 #3383 @KristianNelson
@fsanchez-trigueros thank you for your comments.

I'll try to answer the questions and respond to comments in order.

Generalization:

This package does have potential to be applied elsewhere other than CONUS, and I agree that it would be a huge enhancement which would attract more users. The challenge is that the `gamut` package is rather specific to the data that it pulls in. For example, the crop cover raster has specific values for each cover type, which is then reclassified in the package to GCAM values. This reclassification would have to change if values from non-CONUS crop cover rasters are different.

If a new data package was created for non-CONUS areas, there would be a couple options to make the change successful:

1. Set up the new data package so that it aligns with the data in the CONUS package. This would timely but would help keep the simplicity of the code in `gamut`
2. Lets say data was compiled for Canada, but the data isn't set up the same as the original CONUS data. A "domain" variable in the `count_watershed_teleconnections` function could be set up, which could be changed to "Canada" or "CONUS". Based on this variable, different sets of code could be ran which aligns with either area.

**Installation: Does installation proceed as outlined in the documentation?**

Installation does proceed as outlined. I have not run into the remote installation issue that you encountered, so I appreciate that feedback and I will be sure to provide other installation options so that users can go through different routes if they have issues.

I will also look into the Rtools dependency and make sure that I provide some explanation for users that run into that problem. I'm wondering if it has anything to do with my version of R since I haven't ran into that issue.

**Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).**

There is a paper currently being reviewed that uses the `gamut` package to analyze water reuse in urban cities. When this is published I would like to reference this paper to show how gamut can be applied to the real world.

**References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?**

The list of references is complete and uses the .bib format required by JOSS. The in text citations use proper syntax. If any additional references are used I will be sure to continue using the correct formatting.

**State of the field: Do the authors describe how this software compares to other commonly-used packages?**

This package is similar to other multi-sectoral packages in the sense that it combines several datasets regarding a number of sectors in order to provide useful data within certain domains. An example of this is the `cerf` package which combines multiple datasets to show power plant suitability across the CONUS.

**Who coined the urban teleconnections? What publications talk about it?**

"Urban teleconnections" was originally coined by Seto et. al in this journal article: https://www.researchgate.net/publication/224879342_Urban_Land_Teleconnections_and_Sustainability

I will be sure to put a reference to that and some context for the reader so that they can get some background on the importance of observing urban teleconnections. I do see where there needs to be more background to attract potential users, so I will extend the literature review to provide more depth on this space.

In regards to the ZIP file size, I will look into other options that can provide easier access for users. We knew we wanted to have the data package accessible through an online open-source repository, and Zenodo has always been my first option. There is the option of splitting the data package up into different zip files which would decrease the file size and lessen the computational burden. For example, within the main folder there are subfolders called "water", "land", and "energy" which could be separated. This would only require a small change in the code to make sure the package retrieves them correctly.
