+++ .github/workflows/test_coverage.yml
@@ -1,36 +1,27 @@
-on: [push, pull_request]
+name: test_coverage
 
-name: build
+on:
+  push:
+    branches:
+      - main
 
 jobs:
   build:
-    runs-on: ubuntu-latest
-
-    env:
-      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-
+    runs-on: macOS-latest
+    strategy:
+      matrix:
+        r: ['3.6']
     steps:
       - uses: actions/checkout@v1
       - uses: r-lib/actions/setup-r@master
         with:
-          r-version: '3.6'
+          r-version: ${{ matrix.r }}
       - uses: r-lib/actions/setup-pandoc@master
       - uses: r-lib/actions/setup-tinytex@master
-
-      - name: Cache R packages
-        uses: actions/cache@v1
-        with:
-          path: ${{ env.R_LIBS_USER }}
-          key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }}
-          restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-
-
-      - name: Ubuntu config
-        run: |
-          sudo add-apt-repository ppa:ubuntugis/ppa
-          sudo apt-get update
-          sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev libudunits2-dev libgdal-dev libgeos-dev libproj-dev libmagick++-dev libavfilter-dev
-
-
+      - name: Install dependencies
+        env:
+          R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+        run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'covr'))" -e "remotes::install_deps(dependencies = TRUE)"
       - name: Test coverage
         run: |
           remotes::install_cran("covr")
