+++ .github/workflows/test_coverage.yml
@@ -1,9 +1,6 @@
-name: test_coverage
+on: [push, pull_request]
 
-on:
-  push:
-    branches:
-      - main
+name: build
 
 jobs:
   build:
