+++ .github/workflows/test_coverage.yml
@@ -4,7 +4,7 @@ name: build
 
 jobs:
   build:
-    runs-on: macOS-latest
+    runs-on: ubuntu-latest
     strategy:
       matrix:
         r: ['3.6']
@@ -19,6 +19,11 @@ jobs:
         env:
           R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
         run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'covr'))" -e "remotes::install_deps(dependencies = TRUE)"
+      - name: Ubuntu config
+        run: |
+          sudo add-apt-repository ppa:ubuntugis/ppa
+          sudo apt-get update
+          sudo apt-get install -y libgeos-dev
       - name: Test coverage
         run: |
           remotes::install_cran("covr")
