+++ .github/workflows/test_coverage.yml
@@ -23,7 +23,7 @@ jobs:
         run: |
           sudo add-apt-repository ppa:ubuntugis/ppa
           sudo apt-get update
-          sudo apt-get install -y libgeos-dev
+          sudo apt-get install -y libgeos-dev libcurl4-openssl-dev
       - name: Test coverage
         run: |
           remotes::install_cran("covr")
