+++ paper/paper.bib
@@ -262,3 +262,14 @@
 	url = {http://doi.org/10.5281/zenodo.4662993},
 	Title = {{Geospatial Input Datasets for Urban Teleconnections Analysis}},
 	Year = 2021}
+
+@article{Seto:2012,
+author = {Seto, Karen and Reenberg, Anette and Boone, Christopher and Fragkias, Michail and Haase, Dagmar and Langanke, Tobias and Marcotullio, Peter and Munroe, Darla and Olah, Branislav and Simon, David},
+year = {2012},
+month = {05},
+pages = {7687-92},
+title = {Urban Land Teleconnections and Sustainability},
+volume = {109},
+journal = {Proceedings of the National Academy of Sciences of the United States of America},
+doi = {10.1073/pnas.1117622109}
+}
