+++ README.md
@@ -18,9 +18,7 @@ correlation between human and environmental phenomena that occur a long
 distance apart.
 
 <p align="center">
-
 <img width="800" height="550" src="inst/extdata/gamut_plot.png">
-
 </p>
 
 ## Get Started with `gamut`
@@ -29,6 +27,8 @@ distance apart.
 `devtools` package. From an R prompt, run the command:
 
 ``` r
+install.packages("devtools")
+library(devtools)
 devtools::install_github('IMMM-SFA/gamut')
 library(gamut)
 ```
@@ -41,7 +41,7 @@ the zipped data files to your preferred directory. The table below shows
 all the files used within the `gamut` software package.
 
 | gamut Name       | Sub Folder Location | Full File Path                                                                  | Data Source                                                                                              |
-| :--------------- | :------------------ | :------------------------------------------------------------------------------ | :------------------------------------------------------------------------------------------------------- |
+|:-----------------|:--------------------|:--------------------------------------------------------------------------------|:---------------------------------------------------------------------------------------------------------|
 | watersheds       | water               | water/CWM\_v2\_2/World\_Watershed8.shp                                          | <https://knb.ecoinformatics.org/view/doi%3A10.5063%2FF1J67DWR>                                           |
 | withdrawal       | water               | water/CWM\_v2\_2/Snapped\_Withdrawal\_Points.shp                                | <https://knb.ecoinformatics.org/view/doi%3A10.5063%2FF1J67DWR>                                           |
 | citypoint        | water               | water/CWM\_v2\_2/City\_Centroid.shp                                             | <https://knb.ecoinformatics.org/view/doi%3A10.5063%2FF1J67DWR>                                           |
@@ -80,13 +80,13 @@ learn what each of these variables mean, scroll to the bottom of this
 section and see the table of variables. The result of this function will
 look something like this:
 
-| …1 | city             | city\_population | n\_watersheds | n\_other\_cities | dependent\_city\_pop | watershed\_area\_sqkm | storage\_BCM | yield\_BCM | irr\_cons\_BCM | n\_climate\_zones | n\_transfers\_in | n\_transfers\_out | n\_transfers\_within | n\_hydro\_plants | n\_thermal\_plants | n\_fac\_agcrop | n\_fac\_aglivestock | n\_fac\_cnsmnf | n\_fac\_mining | n\_fac\_oilgas | n\_fac\_total | hydro\_gen\_MWh | thermal\_gen\_MWh | thermal\_cons\_BCM | thermal\_with\_BCM | n\_utilities | n\_ba | n\_crop\_classes | cropland\_fraction | developed\_fraction | ag\_runoff\_max | ag\_runoff\_av\_exgw | ag\_runoff\_av | dev\_runoff\_max | dev\_runoff\_av\_exgw | dev\_runoff\_av | np\_runoff\_max | np\_runoff\_av\_exgw | np\_runoff\_av\_exgw\_unweighted | np\_runoff\_av | n\_economic\_sectors | max\_withdr\_dist\_km | avg\_withdr\_dis\_km | n\_treatment\_plants | watershed\_pop | pop\_cons\_m3sec | av\_fl\_sur\_conc\_pct | av\_fl\_sur\_conc\_pct\_unweighted | av\_ro\_sur\_conc\_pct | av\_fl\_all\_conc\_pct | av\_ro\_all\_conc\_pct | av\_fl\_max\_conc\_pct | av\_ro\_max\_conc\_pct | surface\_contribution\_pct | importance\_of\_worst\_watershed\_pct |
[…]
