+++ README.Rmd
@@ -20,13 +20,25 @@ always_allow_html: yes
 </p>
 
 ## Get Started with `gamut`
-`gamut` can be installed directly from this repository using the R `devtools` package. From an R prompt, run the command:
+
+`gamut` can be installed remotely from the repository using the R `devtools` package. From an R prompt, run the command:
 ```r
 install.packages("devtools")
 library(devtools)
 devtools::install_github('IMMM-SFA/gamut')
 library(gamut)
 ```
+
+If you run into problems with the remote installation, you may also try these other options to install gamut:
+
+1. Save the package file by clicking  [here](https://api.github.com/repos/IMMM-SFA/gamut/tarball/HEAD), then run `install_local()` as shown below:
+```r
+install_local('path/to/package')
+```
+
+2. Clone [the repo][repository] to your computer using `git clone "https://github.com/IMMM-SFA/gamut"`. You can then load this project into your RStudio and install it.
+
+NOTE: Depending on your version of R, you may need to install Rtools to retrieve the package. If you have trouble installing it with `install.packages("Rtools"), you can find the install file [here](https://cran.r-project.org/bin/windows/Rtools/).
 
 ##### Data Files
 
