+++ paper/paper.bib
@@ -273,3 +273,18 @@ volume = {109},
 journal = {Proceedings of the National Academy of Sciences of the United States of America},
 doi = {10.1073/pnas.1117622109}
 }
+
+@article{Rice:2013,
+author = {Rice, Jacelyn and Wutich, Amber and Westerhoff, Paul},
+title = {Assessment of De Facto Wastewater Reuse across the U.S.: Trends between 1980 and 2008},
+journal = {Environmental Science \& Technology},
+volume = {47},
+number = {19},
+pages = {11099-11105},
+year = {2013},
+doi = {10.1021/es402792s},
+note ={PMID: 23957863},
+URL = {https://doi.org/10.1021/es402792s},
+eprint = {https://doi.org/10.1021/es402792s}
+}
+
