+++ DESCRIPTION
@@ -42,7 +42,7 @@ Imports:
     reservoir,
     rgdal,
     rgeos,
-    sf,
+    sf (>= 0.9-4),
     sp,
     spex,
     stringr,
