### 2021-10-07 99c00f4
Add in reference explaining teleconnections, other paper edits
