+++ README.Rmd
@@ -38,11 +38,11 @@ install_local('path/to/package')
 
 2. Clone [the repo][repository] to your computer using `git clone "https://github.com/IMMM-SFA/gamut"`. You can then load this project into your RStudio and install it.
 
-NOTE: Depending on your version of R, you may need to install Rtools to retrieve the package. If you have trouble installing it with `install.packages("Rtools"), you can find the install file [here](https://cran.r-project.org/bin/windows/Rtools/).
+NOTE: Depending on your version of R, you may need to install Rtools to retrieve the package. If you have trouble installing it with `install.packages("Rtools")`, you can find the install file [here](https://cran.r-project.org/bin/windows/Rtools/). Depending on your version of `sf`, also may need to install the package `Rcpp` in order for `gamut` to build correctly. 
 
 ##### Data Files
 
-To download all of the `gamut` input datasets, visit the [Zenodo data repository](https://zenodo.org/record/4662993#.YHTNpehKhPY) and download the zipped data files to your preferred directory. The table below shows all the files used within the `gamut` software package.
+To download all of the `gamut` input datasets, visit the [Zenodo data repository](https://zenodo.org/record/4662993#.YHTNpehKhPY) and download the zipped data files to your preferred directory. Make sure to combine the energy, land, misc, and water folders into a single directory. The table below shows all the files used within the `gamut` software package.
     
 ```{r, echo=FALSE,message=FALSE}
 library(tibble)
@@ -164,6 +164,37 @@ tibble::tribble(
 ) -> variables
 kable(variables)
 ```
+
+### Dependencies
+
+`gamut` relies on functionality from the following R packages: 
+    clisymbols,
+    crayon,
+    dplyr,
+    dams,
+    exactextractr,
+    foreign,
+    geosphere,
+    ggplot2,
+    lwgeom,
+    magrittr,
+    purrr,
+    raster,
+    readxl,
+    reservoir,
+    rgdal,
+    rgeos,
+    sf,
+    sp,
+    spex,
+    stringr,
+    tibble,
+    tidyr,
+    vroom,
+    testthat,
+    knitr,
+    rmarkdown,
+    knitr.
 
 ## Support
 
