+++ paper/paper.md
@@ -37,7 +37,7 @@ Most cities in the United States withdraw surface water to meet public water sup
 
 ![The `gamut` package analyzes urban cities and their watersheds all across the conterminous U.S. As shown in the figure, it can look at characteristics like land use inside watershed boundaries.](gamut_figure.png){ width=95% }
 
-The `gamut` package computes dozens of city-level metrics that inform on the geographical nature of surface water supply catchments and the presence, intensity, and impact of human activities in those catchments. Each city’s watersheds are based on the Urban Water Blueprint [@McDonald:2014], which is enhanced with source contribution estimates as well as river flow and high-resolution runoff [@Nelson:2021]. `gamut` first uses the watershed boundary as a selector for several geospatial layers relating to electricity generation, agriculture, industry and economic developments, and water infrastructure (dams, reservoirs, aqueducts). Similar to other multi-sector packages, `gamut` produces its results from the combination of multiple datasets, which are then used to create new statistics. `gamut` relies heavily on the use of the `st_intersection` function from the [sf](https://cran.r-project.org/web/packages/sf/index.html) package to intersect features and watersheds, which then joins the rest of the information from the geospatial layers. This enables the package to link information such as land use percentages, wastewater discharge, facilities, etc. to a specific city through their drinking water resources. Non-geospatial layers in the form of data tables are joined through city names and municipal IDs. The input layers used in this package have been combined into an open-source dataset and can accessed [here](https://zenodo.org/record/4662993#.YJB9ZLVKhPY) [@Nelson2:2021].
+The `gamut` package computes dozens of city-level metrics that inform on the geographical nature of surface water supply catchments and the presence, intensity, and impact of human activities in those catchments. Each city’s watersheds are based on the Urban Water Blueprint [@McDonald:2014], which is enhanced with source contribution estimates as well as river flow and high-resolution runoff [@Nelson:2021]. `gamut` first uses the watershed boundary as a selector for several geospatial layers relating to electricity generation, agriculture, industry and economic developments, and water infrastructure (dams, reservoirs, aqueducts). Similar to other multi-sector packages, `gamut` produces its results from the combination of multiple datasets, which are then used to create new statistics. `gamut` relies heavily on the use of the `st_intersection` function from the [sf](https://cran.r-project.org/web/packages/sf/index.html) package to intersect features and watersheds, which then joins the rest of the information from the geospatial layers. This enables the package to link information such as land use percentages, wastewater discharge, facilities, etc. to a specific city through their drinking water resources. Non-geospatial layers in the form of data tables are joined through city names and municipal IDs. The input layers used in this package have been combined into an open-source dataset and can accessed [here](https://zenodo.org/record/5554939#.YV9dnNrMJPY) [@Nelson2:2021].
 
 By creating the links between cities and watershed characteristics, the `gamut` package calculates numerous metrics that may be used for multiple types of city level multisector dynamics research. Metrics reported by `gamut` fall into four main categories: geographical characteristics of watersheds (e.g., climate zones, land area, distance from city, hydrology), potential water contamination concentrations (nonpoint and point), withdrawal/consumption of water from other sectors, and presence/intensity of multisectoral land uses. Table 1 shows all of the metrics that are created by this package, descriptions, and units. An R vignette is provided to help users to get started with `gamut` and may be accessed [here](https://github.com/IMMM-SFA/gamut#readme).
 
