+++ README.Rmd
@@ -115,7 +115,7 @@ tibble::tribble(
   "yield_BCM","Combined yield capacity of all the city catchments","billion cubic meters",
   "irr_cons_BCM","Combined water consumption that is used for irrigation with the watersheds","billion cubic meters",
   "n_climate_zones","Number of climate zones that the source watersheds cover","zones",
-  "n_hydro_plants","Number of hydro electric power plants operating within the source watersheds","plants",
+  "n_hydro_plants","Number of hydroelectric power plants operating within the source watersheds","plants",
   "n_thermal_plants","Number of thermal power plants operating within the source watersheds","plants",
   "n_fac_agcrop","Number of agricultural crop facilities within the source watersheds","facilities",
   "n_fac_aglivestock","Number of agicultural livestock facilities within the source watersheds","facilities",
@@ -123,7 +123,7 @@ tibble::tribble(
   "n_fac_mining","Number of mining facilities within the source watersheds","facilities",
   "n_fac_oilgas","Number of oil and gas facilities within the source watersheds","facilities",
   "n_fac_total","Total number of facilities operating within the source watersheds","facilities",
-  "hydro_gen_MWh","Combined hydro electric generation from all the facilities within the source watersheds","megawatt-hours",
+  "hydro_gen_MWh","Combined hydroelectric generation from all the facilities within the source watersheds","megawatt-hours",
   "thermal_gen_MWh","Combined thermal generation from all the facilities within the source watersheds","megawatt-hours",
   "thermal_cons_BCM","Combined water consumption that is used for thermal generation","billion cubic meters",
   "thermal_with_BCM","Combined water withdrawal for thermal generation","billion cubic meters",
