+++ .github/workflows/pkgdown.yml
@@ -2,7 +2,7 @@ on:
   push:
     branches:
       - main
-      - master
+      - dev
 
 name: docs
 
