### 2021-10-08 ee27095
set docs build to ubuntu latest
