+++ .github/workflows/pkgdown.yml
@@ -30,7 +30,7 @@ jobs:
       - name: Install dependencies
         env:
           R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-        run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'covr', 'pkgdown'))" -e "remotes::install_deps(dependencies = TRUE)" 
+        run: Rscript -e "install.packages(c('remotes', 'rcmdcheck', 'covr', 'pkgdown'))" -e "remotes::install_deps(dependencies = TRUE)"
 
       - name: Install package
         run: R CMD INSTALL .
@@ -39,4 +39,5 @@ jobs:
         run: |
           git config --local user.email "actions@github.com"
           git config --local user.name "GitHub Actions"
+          Rscript -e 'install.packages("pkgdown", type = "binary")'
           Rscript -e 'pkgdown::deploy_to_branch(new_process = FALSE)'
