+++ .github/workflows/pkgdown.yml
@@ -39,5 +39,5 @@ jobs:
         run: |
           git config --local user.email "actions@github.com"
           git config --local user.name "GitHub Actions"
-          Rscript -e 'install.packages("pkgdown", type = "binary")'
+          Rscript -e 'install.packages("pkgdown")'
           Rscript -e 'pkgdown::deploy_to_branch(new_process = FALSE)'
