+++ README.md
@@ -58,9 +58,11 @@ order for `gamut` to build correctly.
 ##### Data Files
 
 To download all of the `gamut` input datasets, visit the [Zenodo data
-repository](https://zenodo.org/record/4662993#.YHTNpehKhPY) and download
-the zipped data files to your preferred directory. The table below shows
-all the files used within the `gamut` software package.
+repository](https://zenodo.org/record/5554939#.YV9dnNrMJPY) and download
+the zipped data files to your preferred directory. Make sure to combine
+the energy, land, misc, and water folders into a single directory. The
+table below shows all the files used within the `gamut` software
+package.
 
 | gamut Name       | Sub Folder Location | Full File Path                                                                  | Data Source                                                                                              |
 |:-----------------|:--------------------|:--------------------------------------------------------------------------------|:---------------------------------------------------------------------------------------------------------|
@@ -76,9 +78,9 @@ all the files used within the `gamut` software package.
 | climate          | land                | land/kop\_climate\_classes.tif                                                  | <http://koeppen-geiger.vu-wien.ac.at/present.htm>                                                        |
 | HUC4             | water               | water/USA\_HUC4/huc4\_to\_huc2.shp                                              | <http://prd-tnm.s3-website-us-west-2.amazonaws.com/?prefix=StagedProducts/Hydrography/WBD/National/GDB/> |
 | population       | land                | land/pden2010\_block/pden2010\_60m.tif                                          | <https://www.sciencebase.gov/catalog/item/57753ebee4b07dd077c70868>                                      |
-| runoff           | water               | water/Historical\_Mean\_Runoff/USA\_Mean\_Runoff.tif                            | <https://zenodo.org/record/4315195>                                                                      |
-| nhd\_flow        | water               | water/Watershed\_Flow\_Contributions/UWB\_Intake\_Flows.shp                     | <https://zenodo.org/record/4315195>                                                                      |
-| contributions    | water               | water/Watershed\_Flow\_Contributions/Watershed\_Contributions.csv               | <https://zenodo.org/record/4315195>                                                                      |
+| runoff           | water               | water/UWSCatCH/Historical\_Mean\_Runoff/USA\_Mean\_Runoff.tif                   | <https://zenodo.org/record/4315195>                                                                      |
+| nhd\_flow        | water               | water/UWSCatCH/Watershed\_Flow\_Contributions/UWB\_Intake\_Flows.shp            | <https://zenodo.org/record/4315195>                                                                      |
+| contributions    | water               | water/UWSCatCH/Watershed\_Flow\_Contributions/Watershed\_Contributions.csv      | <https://zenodo.org/record/4315195>                                                                      |
 
 ## Usage
 
@@ -119,59 +121,59 @@ cities’ watersheds.
 The table below shows explanations for each of these variables that are
 created through this function:
 
-| Variable Name                         | Description                                                                             | Units                 |
-|:--------------------------------------|:----------------------------------------------------------------------------------------|:----------------------|
-| city\_population                      | The population of the city being analyzed                                               | people                |
-| n\_watersheds                         | Number of watersheds that city uses to source drinking water                            | watersheds            |
-| n\_other\_cities                      | Number of other cities pulling off the same watersheds                                  | cities                |
-| dependent\_city\_pop                  | Total population of people dependent on that city’s watersheds                          | people                |
-| watershed\_area\_sqkm                 | Combined area of all the source watersheds of a city                                    | square kilometers     |
-| storage\_BCM                          | Combined storage capacity of all the city catchments                                    | billion cubic meters  |
-| yield\_BCM                            | Combined yield capacity of all the city catchments                                      | billion cubic meters  |
[…]
