### 2021-10-11 50ac658
add libfribidi-dev
