+++ README.Rmd
@@ -15,8 +15,10 @@ always_allow_html: yes
 ## Description
 `gamut` is a tool for exploring *teleconnections* between cities of the United States and human activities that occur in their associated water supply catchments. A *teleconnection* is a causal connection or correlation between human and environmental phenomena that occur a long distance apart.
 
+
 <p align="center">
-  <img width="800" height="550" src="inst/extdata/gamut_plot.png">
+<a
+<img src="https://github.com/IMMM-SFA/gamut/blob/main/inst/extdata/gamut_plot.png?raw=true" width="800" height="550"></a> 
 </p>
 
 ## Get Started with `gamut`
@@ -90,15 +92,10 @@ knitr::kable(results)
 
 This table can be used to compare different variables between multiple cities. Below is a graph comparing how much developed land are in cities' watersheds. 
 
-```{r, echo=FALSE,message=FALSE}
-library(ggplot2)
-results %>% 
-  ggplot(aes(x = reorder(city, developed_fraction), y = developed_fraction)) +
-  geom_bar(stat = "identity", fill = "steelblue") +
-  coord_flip() +
-  labs(x = NULL, y = "Fraction of Developed Land") +
-  theme_classic() 
-```
+<p align="center">
+<a
+<img src="https://github.com/IMMM-SFA/gamut/blob/main/inst/extdata/gamut_graph.png?raw=true" ></a> 
+</p>
 
 
 The table below shows explanations for each of these variables that are created through this function:
