+++ R/io.R
@@ -2,7 +2,7 @@
 
 #' Create an sf object from a shapefile
 #'
-#' Create an sf object from a full path to shapefile with file name and extension
+#' @details Create an sf object from a full path to shapefile with file name and extension
 #'
 #' @param shp_path character. A full path to the input shapefile with file name and extension
 #' @param quiet boolean.
@@ -21,7 +21,7 @@ import_shapefile <- function(shp_path, quiet = TRUE, method = "sf") {
 
 #' Create a raster object from a file path
 #'
-#' Create a raster object from a full path to raster file name and extension
+#' @detials Create a raster object from a full path to raster file name and extension
 #'
 #' @param raster_path character. A full path to the input raster file with file name and extension
 #' @return raster object
@@ -35,8 +35,7 @@ import_raster <- function(raster_path) {
 
 #' get_cities
 #'
-#' Read internal data file that specifies all 236 US cities from UWB database.
-#' Provides consistent mapping across cities, intakes and watersheds.
+#' @details Read internal data file that specifies all 236 US cities from UWB database. Provides consistent mapping across cities, intakes and watersheds.
 #' @importFrom vroom vroom cols
 #' @author Sean Turner (sean.turner@pnnl.gov)
 #' @export
@@ -55,7 +54,7 @@ get_cities <- function(){
 
 #' get_crop_mapping
 #'
-#' Read internal data file that specifies the GCAM classification for certain crop types.
+#' @details Read internal data file that specifies the GCAM classification for certain crop types.
 #' @import vroom
 #' @author Kristian Nelson (kristian.nelson@pnnl.gov)
 #' @export
@@ -85,7 +84,7 @@ get_crop_mapping <- function(){
 
 #' get_ucs_power_plants
 #'
-#' Read the UCS database
+#' @details This function reads the UCS power plant database
 #' @param ucs_file_path full path of UCS xlsx file within data_dir
 #' @param method tool used to convert to spatial data; either 'sp' or 'sf'
 #' @importFrom readxl read_xlsx
@@ -169,9 +168,7 @@ get_ucs_power_plants <- function(ucs_file_path,
 
 #' Get raster frequency per class from polygon input areas
 #'
-#' Get the coverage per class of raster values represented in the input raster dataset
-#' when restricted to the input watershed polygons for a target city.
-#'
+#' @details Get the coverage per class of raster values represented in the input raster dataset when restricted to the input watershed polygons for a target city.
 #' @param raster_object character. An object of class RasterLayer.
 #' @param polygon character. A polygon to define spatial boundary of raster value counts (e.g. a given city's watersheds)
 #' @return table of crop types present and their coverage within the polygon area
@@ -204,7 +201,7 @@ get_zonal_data <- function(raster_object, polygon) {
 
 #' Mask raster to polygon
 #'
-#' @details masks a raster file against a chosen polygon.
+#' @details This function masks a raster file against a chosen polygon. This function differs from a normal mask because it makes sure the projections are the same for both layers.
 #' @param raster_object character. An object of class RasterLayer.
 #' @param polygon character. A polygon to define spatial boundary of raster value counts (e.g. a given city's watersheds)
 #' @importFrom sf st_crs st_transform st_as_sf
@@ -324,7 +321,7 @@ reclassify_raster <- function(crop_cover_levels){
 
 #' get_demeter_file
 #' @param irrigation_file_path FUll path to the demeter data file
-#' @details Load demeter file for irrigation and rainfed crop data
[…]
