+++ paper/paper.md
@@ -37,7 +37,7 @@ Most cities in the United States withdraw surface water to meet public water sup
 
 ![The `gamut` package analyzes urban cities and their watersheds all across the conterminous U.S. As shown in the figure, it can look at characteristics like land use inside watershed boundaries.](gamut_figure.png){ width=95% }
 
-The `gamut` package computes dozens of city-level metrics that inform on the geographical nature of surface water supply catchments and the presence, intensity, and impact of human activities in those catchments. Each city’s watersheds are based on the Urban Water Blueprint [@McDonald:2014], which is enhanced with source contribution estimates as well as river flow and high-resolution runoff [@Nelson:2021]. `gamut` first uses the watershed boundary as a selector for several geospatial layers relating to electricity generation, agriculture, industry and economic developments, and water infrastructure (dams, reservoirs, aqueducts). Similar to other multisector packages, `gamut` produces its results from the combination of multiple datasets, which are then used to create new statistics. `gamut` relies heavily on the use of the `st_intersection` function from the [sf](https://cran.r-project.org/web/packages/sf/index.html) package to intersect features and watersheds, which then joins the rest of the information from the geospatial layers. This enables the package to link information such as land use percentages, wastewater discharge, facilities, etc. to a specific city through their drinking water resources. Non-geospatial layers in the form of data tables are joined through city names and municipal IDs. The input layers used in this package have been combined into an open-source dataset and can accessed [here](https://zenodo.org/record/5554939#.YV9dnNrMJPY) [@Nelson2:2021].
+The `gamut` package computes dozens of city-level metrics that inform on the geographical nature of surface water supply catchments and the presence, intensity, and impact of human activities in those catchments. The package cycles through a 3-step process. First, it connects cities to there drinking water resources by using the Urban Water Blueprint dataset [@McDonald:2014]. This dataset is combined with an enhancment dataset of source contribution estimates as well as river flow and high-resolution runoff [@Nelson:2021]. After connecting the cities to the watersheds, the second step is using the watershed boundary as a mask for the input geospatial layers. Additionally, `gamut` relies heavily on the use of the `st_intersection` function from the [sf](https://cran.r-project.org/web/packages/sf/index.html) package to intersect features and watersheds, which joins the information from the geospatial layers. These layers encompass a wide variety of characteristics including land use, land cover, power generation, infrastructure, hydrological data (irrigation, stream flow, runoff), and population data. Similar to other multisector packages, `gamut` produces new data from a combination of multiple datasets, which are then used to create new statistics.  Non-geospatial layers in the form of data tables are joined through city names and municipal IDs. The input layers used in this package have been combined into an open-source dataset and can accessed [here](https://zenodo.org/record/5554939#.YV9dnNrMJPY) [@Nelson2:2021]. The last step in the process is the creation of output metrics, which is done through the calculation of statistics based on the masked input data for each city.
 
 By creating the links between cities and watershed characteristics, the `gamut` package calculates numerous metrics that may be used for multiple types of city level multisector dynamics research. Metrics reported by `gamut` fall into four main categories: geographical characteristics of watersheds (e.g., climate zones, land area, distance from city, hydrology), potential water contamination concentrations (nonpoint and point), withdrawal/consumption of water from other sectors, and presence/intensity of multisectoral land uses. Table 1 shows all of the metrics that are created by this package, descriptions, and units. An R vignette is provided to help users to get started with `gamut` and may be accessed [here](https://immm-sfa.github.io/gamut/).
 
