### 2021-10-13 cbb5a8a
update auths for docs
