+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: gamut
 Type: Package
 Title: Geospatial Analytics for Multisector Urban Teleconnections
 Version: 1.0.0
-Authors@R: 
+Authors@R:
     person(given = "Kristian",
            family = "Nelson",
            role = c("aut", "cre"),
@@ -10,12 +10,12 @@ Authors@R:
            comment = c(ORCID = "0000-0002-6745-167X"))
     person(given = "Sean",
            family = "Turner",
-           role = c("aut", "cre"),
+           role = "aut",
            email = "sean.turner@pnnl.gov",
            comment = c(ORCID = "0000-0003-4400-9800"))
     person(given = "Chris",
            family = "Vernon",
-           role = c("aut", "cre"),
+           role = "aut",
            email = "chris.vernon@pnnl.gov",
            comment = c(ORCID = "0000-0002-3406-6214"))
 Description: gamut is a tool for exploring teleconnections between cities of the United States and human activities that occur in their associated water supply catchments.
@@ -25,7 +25,7 @@ LazyData: true
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.1
 Depends: R (>= 3.5.0)
-Imports: 
+Imports:
     clisymbols,
     crayon,
     dplyr,
@@ -49,7 +49,7 @@ Imports:
     tibble,
     tidyr,
     vroom
-Suggests: 
+Suggests:
     testthat (>= 2.1.0),
     knitr,
     rmarkdown
