+++ DESCRIPTION
@@ -3,21 +3,21 @@ Type: Package
 Title: Geospatial Analytics for Multisector Urban Teleconnections
 Version: 1.0.0
 Authors@R:
-    person(given = "Kristian D.",
+    c(person(given = "Kristian D.",
            family = "Nelson",
            role = c("aut", "cre"),
            email = "kristian.nelson@pnnl.gov",
-           comment = c(ORCID = "0000-0002-6745-167X"))
+           comment = c(ORCID = "0000-0002-6745-167X")),
     person(given = "Sean W.D.",
            family = "Turner",
            role = "aut",
            email = "sean.turner@pnnl.gov",
-           comment = c(ORCID = "0000-0003-4400-9800"))
+           comment = c(ORCID = "0000-0003-4400-9800")),
     person(given = "Chris R.",
            family = "Vernon",
            role = "aut",
            email = "chris.vernon@pnnl.gov",
-           comment = c(ORCID = "0000-0002-3406-6214"))
+           comment = c(ORCID = "0000-0002-3406-6214")))
 Description: gamut is a tool for exploring teleconnections between cities of the United States and human activities that occur in their associated water supply catchments.
 License: BSD_2_clause + file LICENSE
 Encoding: UTF-8
