+++ paper/paper.bib
@@ -199,24 +199,15 @@
 
 @Article{Wickham:2011,
     author = {Hadley Wickham},
-    title = {testthat: Get Started with Testing},
-    journal = {The R Journal},
+    title = {{testthat: Get Started with Testing}},
     year = {2011},
-    volume = {3},
+    journal = {{The R Journal}},
+    doi = {10.32614/RJ-2011-002},
+    url = {https://doi.org/10.32614/RJ-2011-002},
     pages = {5--10},
-    url = {https://journal.r-project.org/archive/2011-1/RJournal_2011-1_Wickham.pdf},
-  }
-
-@InCollection{Xie:2014,
-    booktitle = {Implementing Reproducible Computational Research},
-    editor = {Victoria Stodden and Friedrich Leisch and Roger D. Peng},
-    title = {knitr: A Comprehensive Tool for Reproducible Research in {R}},
-    author = {Yihui Xie},
-    publisher = {Chapman and Hall/CRC},
-    year = {2014},
-    note = {ISBN 978-1466561595},
-    url = {http://www.crcpress.com/product/isbn/9781466561595},
-  }
+    volume = {3},
+    number = {1}
+}
 
 @Book{Xie:2018,
     title = {R Markdown: The Definitive Guide},
@@ -229,15 +220,14 @@
   }
 
 @InCollection{Xie:2014,
-    booktitle = {Implementing Reproducible Computational Research},
-    editor = {Victoria Stodden and Friedrich Leisch and Roger D. Peng},
-    title = {knitr: A Comprehensive Tool for Reproducible Research in {R}},
-    author = {Yihui Xie},
-    publisher = {Chapman and Hall/CRC},
-    year = {2014},
-    note = {ISBN 978-1466561595},
-    url = {http://www.crcpress.com/product/isbn/9781466561595},
-  }
+    title={knitr: a comprehensive tool for reproducible research in R},
+    author={Xie, Yihui},
+    booktitle={Implementing reproducible research},
+    pages={3--31},
+    year={2018},
+    publisher={Chapman and Hall/CRC},
+	doi = {10.1201/9781315373461-1},
+}
 
 @Manual{Wickham:2020,
     title = {devtools: Tools to Make Developing R Packages Easier},
@@ -287,4 +277,3 @@ note ={PMID: 23957863},
 URL = {https://doi.org/10.1021/es402792s},
 eprint = {https://doi.org/10.1021/es402792s}
 }
-
