### 2021-10-13 f9c7e0b
remove dup ref in paper
