+++ paper/paper.bib
@@ -12,7 +12,7 @@
 	Title = {{UWSCatCH: Urban Water Supply Catchment Contributions and Hydrological Statistics for large cities of the conterminous United States.}},
 	Year = 2021}
 
-@misc{Csardi:2017,
+@misc{Csardi2:2017,
     title = {{clisymbols: Unicode Symbols at the R Prompt}},
     author = {Gábor {Csárdi} and Sindre {Sorhus}},
     Year = 2017,
@@ -28,7 +28,7 @@
     url = {https://CRAN.R-project.org/package=crayon},
   }
 
-@Manual{Henry:2020,
+@Manual{Henry2:2020,
     title = {dplyr: A Grammar of Data Manipulation},
     author = {Hadley {Wickham} and Romain {François} and Lionel {
              Henry} and Kirill {Müller}},
@@ -121,7 +121,7 @@
     doi = {10.1016/j.envsoft.2015.11.007},
   }
 
-@Manual{Bivand:2020,
+@Manual{Bivand2:2020,
     title = {rgdal: Bindings for the 'Geospatial' Data Abstraction Library},
     author = {Roger {Bivand} and Tim {Keitt} and Barry {Rowlingson}},
     year = {2020},
@@ -181,7 +181,7 @@
     url = {https://CRAN.R-project.org/package=tibble},
   }
 
-@Manual{Wickham:2020,
+@Manual{Wickham2:2020,
     title = {tidyr: Tidy Messy Data},
     author = {Hadley {Wickham} and Lionel {Henry}},
     year = {2020},
