### 2021-10-13 79fc2ec
fix stringr ref in bib and paper
