+++ paper/paper.bib
@@ -165,7 +165,7 @@
     url = {https://CRAN.R-project.org/package=spex},
   }
 
-@Manual{Wickham:2019,
+@Manual{Wickham2:2019,
     title = {stringr: Simple, Consistent Wrappers for Common String Operations},
     author = {Hadley Wickham},
     year = {2019},
