### 2021-10-20 4cbae59
fix typo and add double brackets to reference titles
