+++ paper/paper.bib
@@ -21,7 +21,7 @@
   }
 
 @Manual{Csardi:2017,
-    title = {crayon: Colored Terminal Output},
+    title = {{crayon: Colored Terminal Output}},
     author = {Gábor Csárdi},
     year = {2017},
     note = {R package version 1.3.4},
@@ -29,7 +29,7 @@
   }
 
 @Manual{Henry2:2020,
-    title = {dplyr: A Grammar of Data Manipulation},
+    title = {{dplyr: A Grammar of Data Manipulation}},
     author = {Hadley {Wickham} and Romain {François} and Lionel {
              Henry} and Kirill {Müller}},
     year = {2020},
@@ -38,16 +38,16 @@
   }
 
 @Manual{Baston:2020,
-    title = {exactextractr: Fast Extraction from Raster Datasets using Polygons},
-    author = {{Daniel Baston}},
+    title = {{exactextractr: Fast Extraction from Raster Datasets using Polygons}},
+    author = {Daniel Baston},
     year = {2020},
     note = {R package version 0.4.0},
     url = {https://CRAN.R-project.org/package=exactextractr},
   }
 
 @Manual{R-Core-Team:2020,
-    title = {foreign: Read Data Stored by 'Minitab', 'S', 'SAS', 'SPSS', 'Stata',
-'Systat', 'Weka', 'dBase', ...},
+    title = {{foreign: Read Data Stored by 'Minitab', 'S', 'SAS', 'SPSS', 'Stata',
+'Systat', 'Weka', 'dBase', ...}},
     author = {{R Core Team}},
     year = {2020},
     note = {R package version 0.8-80},
@@ -55,7 +55,7 @@
   }
 
 @Manual{Hijmans:2019,
-    title = {geosphere: Spherical Trigonometry},
+    title = {{geosphere: Spherical Trigonometry}},
     author = {Robert J. Hijmans},
     year = {2019},
     note = {R package version 1.5-10},
@@ -64,7 +64,7 @@
 
 @Book{Wickham:2016,
     author = {Hadley Wickham},
-    title = {ggplot2: Elegant Graphics for Data Analysis},
+    title = {{ggplot2: Elegant Graphics for Data Analysis}},
     publisher = {Springer-Verlag New York},
     year = {2016},
     isbn = {978-3-319-24277-4},
@@ -72,7 +72,7 @@
   }
 
 @Manual{Pebesma:2020,
-    title = {lwgeom: Bindings to Selected 'liblwgeom' Functions for Simple Features},
+    title = {{lwgeom: Bindings to Selected 'liblwgeom' Functions for Simple Features}},
     author = {Edzer Pebesma},
     year = {2020},
     note = {R package version 0.2-5},
@@ -80,7 +80,7 @@
   }
 
 @Manual{Bache:2014,
-    title = {magrittr: A Forward-Pipe Operator for R},
+    title = {{magrittr: A Forward-Pipe Operator for R}},
     author = {Stefan Milton {Bache} and Hadley {Wickham}},
     year = {2014},
     note = {R package version 1.5},
@@ -88,7 +88,7 @@
   }
 
 @Manual{Henry:2020,
-    title = {purrr: Functional Programming Tools},
+    title = {{purrr: Functional Programming Tools}},
     author = {Lionel {Henry} and Hadley {Wickham}},
     year = {2020},
     note = {R package version 0.3.4},
@@ -96,7 +96,7 @@
   }
 
 @Manual{Hijmans:2020,
-    title = {raster: Geographic Data Analysis and Modeling},
+    title = {{raster: Geographic Data Analysis and Modeling}},
     author = {Robert J. Hijmans},
     year = {2020},
[…]
